fix another recursive dep
