+++ src/core/state/customRPC/index.ts
@@ -1,8 +1,8 @@
 import { Chain } from '@wagmi/chains';
 import create from 'zustand';
 
+import { getDefaultRPC } from '~/core/references';
 import { ChainId } from '~/core/types/chains';
-import { getDefaultRPC } from '~/core/utils/chains';
 
 import { createStore } from '../internal/createStore';
